pyproject.toml
[tool.poetry]
name = "interview-assessment"
version = "1.0.0"
description = "Interview Assessment"
authors = ["Engineering Team <engineering.team@progyny.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
PyMySQL = "^1.0.2"
cryptography = "^3.4.7"
mysql-connector-python = "^8.0.27"
pandas = "^1.3.5"
psycopg2 = "^2.9.2"
SQLAlchemy = "^1.4.28"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
